Admin felület design

+++ PhoneScoutAdmin/Models/FullPhone.cs
@@ -26,7 +26,7 @@ namespace PhoneScoutAdmin.Models
         public int? connectionDualSim { get; set; }
         public int? connectionEsim { get; set; }
         public int? connectionNfc { get; set; }
-        public double? connectionConnectionSpeed { get; set; }
+        public decimal? connectionConnectionSpeed { get; set; }
         public int? connectionJack { get; set; }
 
         public int? sensorsInfrared { get; set; }
